requirements.txt
cognee[ollama,neo4j]>=0.5.0
neo4j>=5.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0
pyvis>=0.3.2
python-dotenv>=1.0.0
pydantic-settings>=2.6.0
httpx>=0.27.0
slowapi>=0.1.9

# Optional: document loader for PDF/DOCX/PPTX/XLSX/HTML
# pip install docling>=2.0.0

# Dev
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.8.0
